feat(bulk): add a "skip reduce" flag to the bulk loader (#9618)
**Description**

This PR add a `--skip_reduce_phase` flag to the bulk loader. When
supplied, the bulk loader stops after the map phase.

Workflow:

```sh
# start a dgraph zero (for uid and ts mgmt)
dgraph bulk --skip_reduce_phase -f data.rdf.gz  -s dql.schema  --tmp tmp
```

Move your tmp folder, or not...

```sh
dgraph bulk --skip_map_phase --tmp tmp --out out
```

Closes #9615 

**Checklist**

- [x] The PR title follows the
[Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/#summary) syntax,
leading
      with `fix:`, `feat:`, `chore:`, `ci:`, etc.
- [x] Code compiles correctly and linting (via trunk) passes locally
- [x] Tests added for new functionality, or regression tests for bug
fixes added as applicable
- [ ] For public APIs, new features, etc., a PR on the
[docs repo](https://github.com/dgraph-io/dgraph-docs) staged and linked
here. This process can
be simplified by going to the [public docs
site](https://docs.dgraph.io/) and clicking the
"Edit this page" button at the bottom of page(s) relevant to your
changes. Ensure that you
indicate in the PR that this is an **unreleased** feature so that it
does not get merged into
      the main docs prematurely.

+++ dgraph/cmd/bulk/run.go
@@ -78,6 +78,8 @@ func init() {
 	flag.Int64("partition_mb", 4, "Pick a partition key every N megabytes of data.")
 	flag.Bool("skip_map_phase", false,
 		"Skip the map phase (assumes that map output files already exist).")
+	flag.Bool("skip_reduce_phase", false,
+		"Skip the reduce phase (stops after map phase completion).")
 	flag.Bool("cleanup_tmp", true,
 		"Clean up the tmp directory after the loader finishes. Setting this to false allows the"+
 			" bulk loader can be re-run while skipping the map phase.")
@@ -150,6 +152,7 @@ func run() {
 		MapBufSize:       uint64(Bulk.Conf.GetInt("mapoutput_mb")),
 		PartitionBufSize: int64(Bulk.Conf.GetInt("partition_mb")),
 		SkipMapPhase:     Bulk.Conf.GetBool("skip_map_phase"),
+		SkipReducePhase:  Bulk.Conf.GetBool("skip_reduce_phase"),
 		CleanupTmp:       Bulk.Conf.GetBool("cleanup_tmp"),
 		NumReducers:      Bulk.Conf.GetInt("reducers"),
 		Version:          Bulk.Conf.GetBool("version"),
@@ -205,27 +208,29 @@ func RunBulkLoader(opt BulkOptions) {
 	}
 	fmt.Printf("Encrypted input: %v; Encrypted output: %v\n", opt.Encrypted, opt.EncryptedOut)
 
-	if opt.SchemaFile == "" {
-		// if only graphql schema is provided, we can generate DQL schema from it.
-		if opt.GqlSchemaFile == "" {
-			fmt.Fprint(os.Stderr, "Schema file must be specified.\n")
-			os.Exit(1)
-		}
-	} else {
-		if !filestore.Exists(opt.SchemaFile) {
-			fmt.Fprintf(os.Stderr, "Schema path(%v) does not exist.\n", opt.SchemaFile)
-			os.Exit(1)
-		}
-	}
-	if opt.DataFiles == "" {
-		fmt.Fprint(os.Stderr, "RDF or JSON file(s) location must be specified.\n")
-		os.Exit(1)
-	} else {
-		fileList := strings.SplitSeq(opt.DataFiles, ",")
-		for file := range fileList {
-			if !filestore.Exists(file) {
-				fmt.Fprintf(os.Stderr, "Data path(%v) does not exist.\n", file)
+	if !opt.SkipMapPhase {
+		if opt.SchemaFile == "" {
+			// if only graphql schema is provided, we can generate DQL schema from it.
+			if opt.GqlSchemaFile == "" {
+				fmt.Fprint(os.Stderr, "Schema file must be specified.\n")
 				os.Exit(1)
+			}
+		} else {
+			if !filestore.Exists(opt.SchemaFile) {
+				fmt.Fprintf(os.Stderr, "Schema path(%v) does not exist.\n", opt.SchemaFile)
+				os.Exit(1)
+			}
+		}
+		if opt.DataFiles == "" {
+			fmt.Fprint(os.Stderr, "RDF or JSON file(s) location must be specified.\n")
+			os.Exit(1)
+		} else {
+			fileList := strings.SplitSeq(opt.DataFiles, ",")
+			for file := range fileList {
+				if !filestore.Exists(file) {
+					fmt.Fprintf(os.Stderr, "Data path(%v) does not exist.\n", file)
+					os.Exit(1)
+				}
 			}
 		}
 	}
@@ -239,6 +244,20 @@ func RunBulkLoader(opt BulkOptions) {
 		fmt.Fprintf(os.Stderr, "Invalid flags: shufflers(%d) should be <= reduce_shards(%d)\n",
 			opt.NumReducers, opt.ReduceShards)
 		os.Exit(1)
+	}
+
+	// Validate skip phase flags
+	if opt.SkipMapPhase && opt.SkipReducePhase {
+		fmt.Fprint(os.Stderr, "Cannot skip both map and reduce phases.\n")
+		os.Exit(1)
+	}
+	if opt.SkipReducePhase {
+		if Bulk.Cmd.Flags().Changed("cleanup_tmp") && opt.CleanupTmp {
+			fmt.Fprint(os.Stderr, "Cannot use --skip_reduce_phase with --cleanup_tmp=true. "+
+				"Temp files must be preserved for the later reduce phase.\n")
+			os.Exit(1)
+		}
+		opt.CleanupTmp = false
 	}
 	if opt.CustomTokenizers != "" {
 		for _, soFile := range strings.Split(opt.CustomTokenizers, ",") {
@@ -267,25 +286,28 @@ func RunBulkLoader(opt BulkOptions) {
 
 	// Make sure it's OK to create or replace the directory specified with the --out option.
 	// It is always OK to create or replace the default output directory.
-	if opt.OutDir != defaultOutDir && !opt.ReplaceOutDir {
-		err := x.IsMissingOrEmptyDir(opt.OutDir)
-		if err == nil {
-			fmt.Fprintf(os.Stderr, "Output directory exists and is not empty."+
-				" Use --replace_out to overwrite it.\n")
-			os.Exit(1)
-		} else if err != x.ErrMissingDir {
-			x.CheckfNoTrace(err)
+	// Skip output directory validation if we're only doing map phase
+	if !opt.SkipReducePhase {
+		if opt.OutDir != defaultOutDir && !opt.ReplaceOutDir {
+			err := x.IsMissingOrEmptyDir(opt.OutDir)
+			if err == nil {
+				fmt.Fprintf(os.Stderr, "Output directory exists and is not empty."+
+					" Use --replace_out to overwrite it.\n")
+				os.Exit(1)
+			} else if err != x.ErrMissingDir {
+				x.CheckfNoTrace(err)
+			}
 		}
-	}
 
-	// Delete and recreate the output dirs to ensure they are empty.
-	x.Check(os.RemoveAll(opt.OutDir))
-	for i := range opt.ReduceShards {
-		dir := filepath.Join(opt.OutDir, strconv.Itoa(i), "p")
-		x.Check(os.MkdirAll(dir, 0700))
-		opt.shardOutputDirs = append(opt.shardOutputDirs, dir)
+		// Delete and recreate the output dirs to ensure they are empty.
+		x.Check(os.RemoveAll(opt.OutDir))
+		for i := range opt.ReduceShards {
+			dir := filepath.Join(opt.OutDir, strconv.Itoa(i), "p")
+			x.Check(os.MkdirAll(dir, 0700))
+			opt.shardOutputDirs = append(opt.shardOutputDirs, dir)
 
-		x.Check(x.WriteGroupIdFile(dir, uint32(i+1)))
+			x.Check(x.WriteGroupIdFile(dir, uint32(i+1)))
+		}
 	}
 
 	// Create a directory just for bulk loader's usage.
@@ -303,10 +325,26 @@ func RunBulkLoader(opt BulkOptions) {
 	x.Check(os.MkdirAll(bufDir, 0700))
 	defer os.RemoveAll(bufDir)
 
-	loader := newLoader(&opt)
-
 	const bulkMetaFilename = "bulk.meta"
 	bulkMetaPath := filepath.Join(opt.TmpDir, bulkMetaFilename)
+	const writeTsFilename = "write.ts"
+	writeTsPath := filepath.Join(opt.TmpDir, writeTsFilename)
+
+	var precomputedWriteTs uint64
+	if opt.SkipMapPhase {
+		writeTsData, err := os.ReadFile(writeTsPath)
+		if err != nil {
+			fmt.Fprintln(os.Stderr, "Error reading write timestamp file; was --skip_reduce_phase used in the map run?")
+			os.Exit(1)
+		}
+		precomputedWriteTs, err = strconv.ParseUint(strings.TrimSpace(string(writeTsData)), 10, 64)
+		if err != nil {
+			fmt.Fprintln(os.Stderr, "Error parsing write timestamp file")
+			os.Exit(1)
+		}
+	}
+
+	loader := newLoader(&opt, precomputedWriteTs)
 
 	if opt.SkipMapPhase {
 		bulkMetaData, err := os.ReadFile(bulkMetaPath)
@@ -343,7 +381,20 @@ func RunBulkLoader(opt BulkOptions) {
 			fmt.Fprintln(os.Stderr, "Error writing to bulk meta file")
 			os.Exit(1)
 		}
+		if err = os.WriteFile(writeTsPath, []byte(strconv.FormatUint(loader.writeTs, 10)), 0600); err != nil {
+			fmt.Fprintln(os.Stderr, "Error writing write timestamp file")
+			os.Exit(1)
+		}
 	}
+
+	if opt.SkipReducePhase {
+		fmt.Println("Skipping reduce phase. Map phase completed successfully.")
+		fmt.Println("Temp files preserved for later reduce phase processing.")
+		// Don't call cleanup() to preserve temp files
+		loader.prog.endSummary()
+		return
+	}
+
 	loader.reduceStage()
 	loader.writeSchema()
 	loader.cleanup()
